dateRangeInput

dateRangeInput

dr_dateRangeInput(inputId, label, start = NULL, end = NULL,
  min = NULL, max = NULL, format = "yyyy-mm-dd",
  startview = "month", weekstart = 0, language = "fr",
  separator = " et ", width = NULL, autoclose = TRUE)

Arguments

inputId
The input slot that will be used to access the value.

label
Display label for the control, or NULL for no label.

start
The initial start date. Either a Date object, or a string in yyyy-mm-dd format. If NULL (the default), will use the current date in the client's time zone.

end
The initial end date. Either a Date object, or a string in yyyy-mm-dd format. If NULL (the default), will use the current date in the client's time zone.

min
The minimum allowed date. Either a Date object, or a string in yyyy-mm-dd format.

max
The maximum allowed date. Either a Date object, or a string in yyyy-mm-dd format.

format
The format of the date to display in the browser. Defaults to "yyyy-mm-dd".

startview
The date range shown when the input object is first clicked. Can be "month" (the default), "year", or "decade".

weekstart
Which day is the start of the week. Should be an integer from 0 (Sunday) to 6 (Saturday).

language
The language used for month and day names. Default is "en". Other valid values include "ar", "az", "bg", "bs", "ca", "cs", "cy", "da", "de", "el", "en-AU", "en-GB", "eo", "es", "et", "eu", "fa", "fi", "fo", "fr-CH", "fr", "gl", "he", "hr", "hu", "hy", "id", "is", "it-CH", "it", "ja", "ka", "kh", "kk", "ko", "kr", "lt", "lv", "me", "mk", "mn", "ms", "nb", "nl-BE", "nl", "no", "pl", "pt-BR", "pt", "ro", "rs-latin", "rs", "ru", "sk", "sl", "sq", "sr-latin", "sr", "sv", "sw", "th", "tr", "uk", "vi", "zh-CN", and "zh-TW".

separator
String to display between the start and end input boxes.

width
The width of the input, e.g. '400px', or '100%'; see validateCssUnit.

autoclose
Whether or not to close the datepicker immediately when a date is selected.